프로덕트 페이지 구현

+++ src/services/localStorageWorker.js
@@ -1,4 +1,4 @@
-import { localStorageHelper as LSHelper } from "utils/localStorageHelper";
+import { localStorageHelper as LSHelper } from 'utils/localStorageHelper';
 
 // TODO: 카드를 클릭하면 recentList에 등록해야한다.
 export function setItemToRecentList(item) {
@@ -8,7 +8,7 @@ export function setItemToRecentList(item) {
     recentList.splice(existItemIndex, 1);
   }
   recentList.push(item);
-  LSHelper.setItem("recentList", {
+  LSHelper.setItem('recentList', {
     date: new Date().toLocaleDateString(),
     data: recentList,
   });
@@ -16,31 +16,32 @@ export function setItemToRecentList(item) {
 // TODO: recentList에 등록할 때 원래 있었는지 체크한다.
 export function checkAlreadyExistInRecentList(id) {
   const recentList = getRecentList();
-  return recentList.findIndex((item) => item.id === id);
+  return recentList.findIndex(item => item.id === id);
 }
 // TODO: NotInterested에 등록할 때 원래 있었는지 체크한다.
 export function checkAlreadyExistInNotInterested(id) {
-  const recentList = getNotInterested();
-  return recentList.findIndex((item) => item.id === id);
+  const notInterested = getNotInterested();
+  return notInterested.includes(id);
 }
 // TODO: 관심없음 버튼을 클릭하면 notInterested에 등록해야한다.
 export function setItemToNotInterested(id) {
   const existItemIndex = checkAlreadyExistInNotInterested(id);
 
-  existItemIndex === -1 &&
-    LSHelper.setItem("notInterested", {
+  if (!existItemIndex) {
+    LSHelper.setItem('notInterested', {
       date: new Date().toLocaleDateString(),
-      data: getNotInterested().push(id),
+      data: [...getNotInterested(), id],
     });
+  }
 }
 // TODO: 랜덤 상품 조회 버튼 클릭시 recentList에서 가져온 후 현재 id객체 제외
 export function exceptNowItem(id) {
-  return getRecentList().filter((data) => data.id !== id);
+  return getRecentList().filter(data => data.id !== id);
 }
 // TODO: 조회목록 페이지에 들어가면 recentList 전부를 리턴해줘야한다.
 export function getRecentList() {
-  const recentList = LSHelper.getItem("recentList") || {
-    date: "",
+  const recentList = LSHelper.getItem('recentList') ?? {
+    date: '',
     data: [],
   };
   return recentList.date === new Date().toLocaleDateString()
@@ -48,7 +49,10 @@ export function getRecentList() {
     : [];
 }
 export function getNotInterested() {
-  const notInterestedList = LSHelper.getItem("notInterested");
+  const notInterestedList = LSHelper.getItem('notInterested') ?? {
+    date: '',
+    data: [],
+  };
   return notInterestedList.date === new Date().toLocaleDateString()
     ? notInterestedList.data
     : [];
